Moved StorageType back to storage package and created ResourceType.
