+++ src/main/java/one/stayfocused/backend/storage/ResourceType.java
@@ -1,0 +1,5 @@
+package one.stayfocused.backend.storage;
+
+public enum ResourceType {
+    AVATAR
+}
