+++ src/main/java/one/stayfocused/backend/storage/StorageType.java
@@ -1,4 +1,4 @@
-package one.stayfocused.backend.service.avatar;
+package one.stayfocused.backend.storage;
 
 public enum StorageType {
     LOCAL, CLOUDINARY
